chore: abstract end crystal behavior to its own file
